requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cycler==0.11.0
Django==4.0.3
flake8==4.0.1
fonttools==4.30.0
idna==3.3
joblib==1.1.0
kiwisolver==1.4.0
matplotlib==3.5.1
mccabe==0.6.1
nltk==3.7
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
regex==2022.3.2
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
textblob==0.17.1
tqdm==4.63.0
tweepy==4.6.0
tzdata==2021.5
urllib3==1.26.8
